+++ src/models/ReproConfig.ts
@@ -18,6 +18,16 @@ export interface ReproConfig {
   allCrashes?: ReproStepsMode;
 
   /**
+   * Repro steps mode for Crash Reporting.
+   *
+   * @default ReproStepsMode.enabledWithNoScreenshots
+   * @deprecated  Please migrate to the new allCrashes flag
+   *
+   * When this is set, `anr`, `appHangs`,`fatalCrash`,`nonFatalCrash`,`forceRestart`, and `outOfMemory` will be ignored.
+   */
+  crash?: ReproStepsMode;
+
+  /**
    * Repro steps mode for Session Replay.
    *
    * @default ReproStepsMode.enabled
@@ -27,41 +37,41 @@ export interface ReproConfig {
   /**
    * Repro steps mode for ANR errors on android.
    *
-   * @default ReproStepsMode.enabled
+   * @default ReproStepsMode.enabledWithNoScreenshots
    */
   anr?: ReproStepsMode;
 
   /**
    * Repro steps mode for AppHangs report.
    *
-   * @default ReproStepsMode.enabled
+   * @default ReproStepsMode.enabledWithNoScreenshots
    */
   appHangs?: ReproStepsMode;
 
   /**
    * Repro steps mode for Fatal crashes.
    *
-   * @default ReproStepsMode.enabled
+   * @default ReproStepsMode.enabledWithNoScreenshots
    */
   fatalCrash?: ReproStepsMode;
   /**
    * Repro steps mode for non-fatal crashes.
    *
-   * @default ReproStepsMode.enabled
+   * @default ReproStepsMode.enabledWithNoScreenshots
    */
   nonFatalCrash?: ReproStepsMode;
 
   /**
    * Repro steps mode for force restart crashes.
    *
-   * @default ReproStepsMode.enabled
+   * @default ReproStepsMode.enabledWithNoScreenshots
    */
   forceRestart?: ReproStepsMode;
 
   /**
    * Repro steps mode for outOfMemory crashes on ios.
    *
-   * @default ReproStepsMode.enabled
+   * @default ReproStepsMode.enabledWithNoScreenshots
    */
   outOfMemory?: ReproStepsMode;